Added Moko Resources Test
- Fixed MastodonApiTests.kt and made it use .json files as resources.
