+++ data/network/src/commonMain/kotlin/social/androiddev/common/network/MastodonApiImpl.kt
@@ -4,6 +4,7 @@ import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.plugins.ResponseException
 import io.ktor.client.request.get
+import io.ktor.serialization.*
 import kotlinx.serialization.SerializationException
 import social.androiddev.common.network.model.Instance
 
@@ -23,6 +24,8 @@ class MastodonApiImpl(
         } catch (exception: SerializationException) {
             Result.failure(exception = exception)
         } catch (exception: ResponseException) {
+            Result.failure(exception = exception)
+        } catch (exception: JsonConvertException) {
             Result.failure(exception = exception)
         }
     }
