+++ data/network/src/commonTest/resources/MR/files/response_instance_invalid.json
